Updated as per comment

+++ app/Services/Importers/AirportImporter.php
@@ -28,15 +28,15 @@ class AirportImporter extends BaseImporter
         $count = 0;
         $rows = $this->db->readRows($this->table, $this->idField, $start, $fields);
         foreach ($rows as $row) {
-            $ground_handling_cost = $row->ground_handling_cost;
-            $fuel_jetA_cost = $row->fuel_jeta_cost;
+            $ground_handling_cost = (float) $row->ground_handling_cost;
+            $fuel_jetA_cost = (float) $row->fuel_jeta_cost;
 
             if ($ground_handling_cost === null && $ground_handling_cost !== 0) {
-                $ground_handling_cost = setting('general.default_ground_handling_cost');
+                $ground_handling_cost = (float) setting('general.default_ground_handling_cost');
             }
 
             if ($fuel_jetA_cost === null && $fuel_jetA_cost !== 0) {
-                $fuel_jetA_cost = setting('general.default_jetA_fuel_cost');
+                $fuel_jetA_cost = (float) setting('general.default_jetA_fuel_cost');
             }
 
             $attrs = [
@@ -47,8 +47,8 @@ class AirportImporter extends BaseImporter
                 'lat'                  => $row->lat,
                 'lon'                  => $row->lng,
                 'hub'                  => $row->hub,
-                'ground_handling_cost' => (float) $ground_handling_cost,
-                'fuel_jeta_cost'       => (float) $fuel_jetA_cost,
+                'ground_handling_cost' => $ground_handling_cost,
+                'fuel_jeta_cost'       => $fuel_jetA_cost,
             ];
 
             $w = ['id' => $attrs['id']];
